02-bridge/requirements.txt
influxdb-client
paho-mqtt==1.5.1
